Assistant checkpoint: Cambiar medidas de llantas a decimal y quitar validaciones
Assistant generated file changes:
- tuco.Clases/DTOs/Inventario/LlantaDTO.cs: Cambiar tipos de Ancho y Perfil a decimal para permitir valores como 250.58
- GestionLlantera.Web/wwwroot/js/views/Inventario/editarProducto.js: Quitar todas las restricciones de validación de medidas de llantas
- tuco.Clases/Models/Llanta.cs: Cambiar tipos de Ancho y Perfil a decimal en el modelo principal
- API/Controllers/InventarioController.cs: Actualizar referencias a medidas de llantas para usar decimal, Actualizar parámetros del método para usar decimal

---

User prompt:

Me refiero  a las medidas de las llantas
yo coloco por ejemplo 250.58  y no se muestra

+++ API/Controllers/InventarioController.cs
@@ -1365,8 +1365,8 @@ namespace API.Controllers
             [FromQuery] int tamano = 12,
             [FromQuery] string? busqueda = null,
             [FromQuery] string? marca = null,
-            [FromQuery] int? ancho = null,
-            [FromQuery] int? perfil = null,
+            [FromQuery] decimal? ancho = null,
+            [FromQuery] decimal? perfil = null,
             [FromQuery] string? diametro = null)
         {
             try
